yolo_series/yolo11.py
# Import necessary libraries
import torch
import torchvision
from torch.utils.data import DataLoader, Dataset
from torchvision.models.detection.faster_rcnn import FastRCNNPredictor
import pandas as pd
import numpy as np
import cv2
import os
import random
import matplotlib.pyplot as plt
from sklearn.model_selection import train_test_split
from sklearn.metrics import average_precision_score
from PIL import Image, ImageDraw
from ultralytics import YOLO

SEED = 42
torch.manual_seed(SEED)
np.random.seed(SEED)
random.seed(SEED)
torch.cuda.manual_seed(SEED)
os.environ['PYTHONHASHSEED'] = str(SEED)
torch.backends.cudnn.deterministic = True
torch.backends.cudnn.benchmark = False

"""
YOLOv11 evaluation stub (Ultralytics).

- Uses environment variables to avoid personal absolute paths:
  - YOLO_DATA_CONFIG: data.yaml path
  - YOLO_WEIGHTS: trained weights path
"""

# Provide config paths via environment variables
data_config = os.getenv("YOLO_DATA_CONFIG", "path/to/data.yaml")

# Example: build from YAML and transfer weights
# model = YOLO("yolo11n.yaml").load("yolo11x.pt")
model = YOLO("yolo11x.pt")

# Example training (uncomment to use):
# results = model.train(
#     data=data_config,
#     epochs=150,
#     batch=16,
#     imgsz=[256, 1600],
#     save_period=1,
#     rect=True,
#     seed=SEED,
# )

# Load trained model weights from env variable
model = YOLO(os.getenv("YOLO_WEIGHTS", "path/to/best.pt"))

# Evaluate model performance
metrics = model.val(data=data_config, imgsz=[256, 1600], rect=True, save_json=True)
# print(metrics)


