requirements.txt
addict==2.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
astor==0.8.1
autograd==1.7.0
autograd-gamma==0.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.1.1
crcmod==1.7
cryptography==44.0.1
cycler==0.12.1
et_xmlfile==2.0.0
filelock==3.14.0
fonttools==4.56.0
formulaic==1.1.1
fsspec==2025.2.0
graphlib_backport==1.1.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
interface-meta==1.3.0
Jinja2==3.1.5
jmespath==0.10.0
joblib==1.4.2
kiwisolver==1.4.7
lifelines==0.27.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mmcv==2.1.0
# Use a pinned mmdetection release instead of a local editable path.
mmdet==3.3.0
mmengine==0.10.6
model-index==0.1.11
mpmath==1.3.0
networkx==3.1
ninja==1.11.1.3
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
opencv-python==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openpyxl==3.1.5
openxlab==0.1.2
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==2.0.3
patsy==1.0.1
pillow==10.4.0
platformdirs==4.3.6
prettytable==3.11.0
psutil==7.0.0
py-cpuinfo==9.0.0
pycocotools==2.0.7
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.19.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2023.4
PyYAML @ file:///croot/pyyaml_1728657952215/work
requests==2.28.2
rich==13.4.2
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
shapely==2.0.7
six==1.17.0
statsmodels==0.14.1
sympy==1.13.3
tabulate==0.9.0
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tomli==2.2.1
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.65.2
triton==3.0.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.107
ultralytics-thop==2.0.14
urllib3==1.26.20
wcwidth==0.2.13
wrapt==1.17.2
yapf==0.43.0
zipp==3.20.2
